// src/app/common/interfaces/broker-form-data.ts
export interface BrokerFormData {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  urlPath: string;
  mobilePhoneNumber: string;
  companyName: string;
  companyPhoneNumber: string;
  TAXID: string;
  disbursementAccountInfo: string;
  websiteURL: string;
  brandingType: string;
  logoImageFile: File;
  eSignature: string;
  businessWebsiteURL: string;
  backEndFee: number;
  additionalDocumentFile: File;
}